[WeightedSumLOP] Work around a g++-4.3 bug.
Apparently g++-4.3 does not like default arguments of the form "... =
FieldVector<K, size>(1)".  It will assume the the default argument ends ad the
"," and that "size" is the type in the declaration of a new function
parameter.  The way around this is to write "... = (FieldVector<K, size>(1))"
or to use a typedef.

Imported from SVN: r607 at trunk