[TypeTree] Fix tree traversal when using rvalue references
During tree traversal, both the tree and the visitor
can be either const or non-const. While this is handled
by special-casing for all four possibilities in the
standard case, rvalue references allow for a much
more elegant solution.

In order for this to work, the type of the Node has
to be a template parameter of the apply() functions
and not the enclosing class. This patch moves the
template parameter there.

Moreover, it also moves the common base class for
traversing composite and power nodes out of the
anonymous namespace to fix a gcc warning.

Imported from SVN: r918 at branches/review_2011