[TypeTree] Enhance GridFunctionSpace::dataHandleGlobalIndices()
The algorithm for calculating the global indices
that belong to a given entity, implemented in
dataHandleGlobalIndices(), is not very efficient
for composed GFSs, as a lot of data is copied repeatedly.
This change to the leaf GFS implementation will allow
for a more efficient implementation. It extends the
interface of the method with a parameter that specifies
an offset into the global container and a bool parameter
for deciding whether or not to resize that container.

As these additional parameters are implementation details,
the original methods are retained and forward to the
new, enhanced methods.

Imported from SVN: r955 at branches/review_2011