[TypeTree] Add fixed capacity stack
This patch adds a stack with fixed capacity that
is allocated on the stack. The main purpose of this
class is to avoid dynamic allocations when iterating
over TypeTrees with a dynamic TreePath by storing
the TreePath in such a fixed capacity stack. The
storage of the stack is allocated in the FixedCapacityStack
class, while its base class FixedCapacityStackView
is a light-weight proxy object that shares its data
with the underlying FixedCapacityStack. The main
reason for this implementation was again the DynamicTreePath
that is passed around as an object and thus frequently
copied. In order to avoid templating the view on the
stack capacity, the stack implements an erasure scheme.

Imported from SVN: r1074 at trunk