[TypeTree] Replace Wrapping*Transformation with Generic*Transformation
It is sometimes necessary to extract type or runtime
information from the transformation. The new
Generic*Transformation descriptors support this by
passing the transformation to the node constructor
and by passing the transformation type to the template structs
responsible for constructing the type of the transformed
Power/CompositeNodes.

Imported from SVN: r1162 at trunk