[GlobalFunctionSpace,LocalFunctionSpace]
Change GFS::globalIndices, to take an iterator range instead of
std::vector. Thsi allows us to change the LFS::bind TMP to write into
the gloabl vector directly.

This was tested with adapt.cc. The percentage needed for a bind went
down from 17% to 7% of the total runtime. Now the 34% of the bind are
spent in UGGrid, which was 21% before. This means to actual bind
became ~ three times faster (4% of the total time instead of 14%).


Imported from SVN: r1186 at trunk