[Ordering] Fix visitors for new Ordering implementations
The visitors in the new Ordering implementations use
the pre() method to act on the direct children of
the current ordering. Unfortunately, this does not
work as

1) pre() will not get called for tree leafs - for
   those, leaf(), will be called instead
2) pre() also gets called on the root of the tree,
   i.e. the current ordering itself. This does in
   several cases cause an infinite recursion

The correct method for looking at all children and
only the children of the node is beforeChild().

Imported from SVN: r1293 at trunk