[Local Containers] Add weighted accumulation views to LocalVector and LocalMatrix
It is now possible to obtain views for LocalVector
and LocalMatrix that only allow accumulation and
take into account a weight during the process.

The views can be obtained by calling
weightedAccumulationView() and mainly have the following
interface:

accumulate() - adds value to entry, taking weight into account
weight() - get weight
rawAccumulate() - adds value to entry unmodified

The last two methods together can be used to move
the weight multiplication out of the loop over the
degrees of freedom.

For backwards compatibility, it is also still possible
to access the containers using operator[] or operator()
as before, but the views will return proxies that
only support operator+=() and operator-=() which
take the weight into account.

Imported from SVN: r1305 at trunk