[Local Containers] Add compatibility methods to old containers to ensure compatibility between converted local operators and old GridOperatorSpace
Local operators that have been converted to use
accumulate() / rawAccumulate() for writing to the
residual / jacobian will not work when used with
the traditional GridOperatorSpace, because in this
case, they directly get the LocalVector / LocalMatrix.
In order to ease the transition, this patch adds the
required methods to those two classes, but immediately
deprecates them to make sure we won't miss any place
in the infrastructure where we hand out the bare
class instead of a view.

Imported from SVN: r1314 at trunk