Began some module documentation for the grid operator stuff.
Also added the method assemblerCell to the grid operator interface,
which receives only the current cell in advance of any binding of the
local function spaces. This method was necessary to enforce a correct
behaviour for the nonoverlapping assembling but may also be used for
caching.


Imported from SVN: r1322 at trunk