[LocalVector] Add helper methods for accessing the BaseContainer from LocalVector
Calling accessBaseContainer() with a LocalVector
returns its BaseContainer; other types are returned
unmodified. This is necessary for the matrix-vector
multiplication in LocalMatrix and for the vread(),
vwrite() and vadd() methods in LocalFunctionSpace,
as those get passed both LocalVectors and regular
std::vectors in different places.

Imported from SVN: r1340 at trunk