[LocalFunctionSpace] BUGFIX: Make sure copied LocalFunctionSpaces have consistent internal state
Rebecca discovered a nasty bug in the LocalFunctionSpace
while working on the dnaplfv code from the howto:

LocalFunctionSpaces contain a vector for storing
the local -> global mapping, but they all share
the vector of the root node. For this purpose, they
get initialized with this pointer in the LocalfunctionSpace
constructor.

This scheme breaks when copying a LocalFunctionSpace
with the default copy constructor: The root node will
have its own, new vector for storing the mapping, but
all pointers still reference the vector of the original
root node. That can create very subtle problems, as
there is never an invalid memory access. In this case,
the problem showed up in conjunction with the
GridFunctionSubSpace, where the upMap() stage was
not done correctly.

This patch fixes the problem by adding custom copy
constructors, but we should probably also think
about the desired semantics: Should it be possible
to copy a LocalFunctionSpace?

Imported from SVN: r1374 at trunk