[Dirichlet Constraints] Avoid unnecessary loop over element DOFs when intersection is not Dirichlet
In the Dirichlet constraints code, an intersectino
gets tested exactly once at its midpoint. Even if
it is determined to not be constrained, the code
still loops over all coefficients of the element
to check whether they would be affected by a Dirichlet
constraint. This patch removes that loop, mainly to
fix a problem in the MultiDomain code.

Imported from SVN: r1387 at trunk