[OneStepGridOperator] Correctly apply constraints in OneStepGridOperator
In interpolate(), the OneStepGridOperator copies
the unconstrained DOFS of the old solution to the
new solution. For this purpose, it uses the constraints
stored in its LocalAssembler.

Pass the constraints of the first child GridOperator
to the base class of the one-step LocalAssembler to
make sure the correct constraints are applied.

Imported from SVN: r1400 at trunk