[GridFunctionToFunctionAdapter] Do not assume dim == dimWorld
By just reexporting the traits of the GridFunction
(which takes element-local, dim-dimensional) coordinates,
the adapter breaks when used in conjunction with
a GridFunction on a GridView with dimWorld > dim.

Fixed by extracting the correct domain information
from the GridView (just as the GridFunctionTraits
already do).

Imported from SVN: r1487 at trunk