[Ordering] Change construction method for composite orderings to use TypeTree transformations
Christian discovered that the ordering construction for composite GFS
fails if variadic templates are not available. This is caused by the
current implementation trying to simply extract the ordering of all
children during CompositeGFS construction and passing those orderings
on to the new composite Ordering object.
This will not work if composite nodes are implemented without variadic
templates, as in that case, there will always be 10 children, but some
of them will be of type EmptyNode and should be ignored. I did some
experimenting, but the most portable way is to just use the TypeTree
transformation mechanism for the construction of the Ordering. For this
purpose, I have extended that algorithm to make it possible to stop the
recursive transformation process at non-leaf nodes, so that only the
root node (the new GFS under construction) will actually be transformed;
its children will just return their pre-built ordering objects and stop
the recursion.

Right now, this method is only used for the CompositeGridFunctionSpace,
but it could also easily be extended to the PowerGridFunctionSpace.

Imported from SVN: r1523 at trunk