[Ordering] Store child orderings as const
Sanitizing the access to the Ordering object of
GridFunctionSpaces has made it impossible to access
a writable reference/pointer to the Ordering of a
GFS during tree transformation, as the tree is always
passed in as a const reference.

On the other hand, an ordering probably should not
muck around with the data of its children in any case,
so this patch simply makes all children of Power/-
CompositeOrderings const.

Imported from SVN: r1527 at trunk