[ComponentBlockwiseOrdering] Fix backwards compatibility with GridFunctionSpaceComponentBlockwiseMapper
As ComponentBlockwiseOrderingTag is a template, we
cannot use a typedef for creating backwards compatibility
with the old GridFunctionSpaceComponentBlockwiseMapper,
but have to use inheritance instead. Unfortunately,
the derived class does not match the template specializations
used in the implementation of the ordering, causing
a compilation failure if code uses the older type to
specify the ordering.

Fixed by adding new lookupNodeTransformation() specializations
for the old version and nesting a normalized tag inside
the actual tag, which can then be extracted in the ordering
implementation.

Imported from SVN: r1542 at trunk