[LA backends] Extend LA backend interface to support batch access and port assembler to use new interface
Due to its design, PETSc is horribly slow when matrix
entries are updated one at a time. There is a batch
interface, but that is not usable through the current
LA backend design, as it requires to retain a certain
amount of state during a batch update. Moreover, PETSc
requires some finalizing calls after all updates (batch
or not) have been made.

This patch extends the PDELab LA backend API in a way
that supports using PETSc:

- Two new static methods on the Backend, flush() and
  finalize(). The first method is supposed to process
  all pending updates issued to the matrix, but still
  allow for more updates after that call. The second
  method should be called right before the end of the
  assembly and is supposed to leave the matrix in a state
  that can be used with the respective LA package.

- The backend contains a class called Accessor, which
  allows for batch updates during the assembly. For this
  purpose, the accessor takes the matrix as well as bound
  LocalFunctionSpaces for both test and trial space.
  The accessor can then be used to update the submatrix
  spanned by those two local spaces. Access is using LOCAL
  indices! Entries can be accessed using get(), set() and
  add(). The batch is executed as soon as the accessor object
  gets destructed.

Imported from SVN: r1607 at trunk