[PETSc] Add a preliminary autoconf test for PETSc
WARNING: For this test I just took the ParMetis test
from dune-common and tweaked it until it worked for
me, someone with a decent knowledge of autotools should
definitely take a look at it. In particular, the test
currently does not correctly handle the dependencies on
X11 and MPI. Moreover, it requires PETSc to be compiled
with C++ support.

Imported from SVN: r1615 at trunk