[LocalFunctionSpace] Integrate new MultiIndex
LocalFunctionSpace will now, in addition to global indices, also
provide multiindices for its degrees of freedom by way of the new
multiIndex() method.

As the MultiIndex class is templated on the maximum length of the
index tuple (which in turn corresponds to the GridFunctionSpace tree
depth), the LocalFunctionSpace nodes need to be templates on the type
of the MultiIndex. Propagating that information required turning
gfs_to_lfs into a template and writing custom transformation templates
for non-leaf GridFunctionSpaces. The patch adjusts all occurrences
within PDELab, but if someone has written a custom GridFunctionSpace,
the transformation registration will need to be adapted.

Imported from SVN: r1655 at trunk