[Ordering] Add new class DirectLeafLocalOrdering
DirectLeafLocalOrdering is a special version of a
local ordering designed for the case where the
GridViewOrdering is attached to a leaf GridFunctionSpace.

In order to avoid efficiency problems for this common
scenario, both the GridViewOrdering and the associated
LocalOrdering differ from the default implementation
to reduce their memory footprint.

Imported from SVN: r1748 at branches/new_dof_handling