[GridFunctionSpace] Port LocalAssembler::add_entry() method to index caches and new Pattern structure
add_entry() is responsible for adding a matrix entry
specified in local indices of two LocalFunctionSpaces
to the global pattern, potentially adding additional
entries due to constraints on the row and/or the
column.
This is a total rewrite of that method, using the
more convenient constraints interface exposed by
the LFS index caches.

Imported from SVN: r1808 at branches/new_dof_handling