[ISTLVectorBackend] Make sure the global <-> local transfer methods also work with plain std::vector
The current implementation of the transfer methods
assumes that the methods are always passed a LocalVector
(or a view on one), but the original methods on the
LocalFunctionSpace were called with a plain std::vector
in some places (interpolation, in particular).

Fix the problem by using the accessBaseContainer()
function from localvector.hh, which will automatically
do the right thing.

Imported from SVN: r1815 at branches/new_dof_handling