[Constraints] Update constraints-based vector manipulation functions to new API
Make the constraints-based vector manipulation functions
work with the new DOF- and ContainerIndex. Note that this
required an API change: All methods must now be passed the
underlying GridFunctionSpace as their first parameter, as
we need access to its ordering so we can translate DOFIndices
to ContainerIndices.

The current implementation is very simplistic, especially
the functions operating on nonconstrained DOFs need to be
optimized. Unfortunately, that task is rather involved, as
currently we cannot iterate over the vector container. Even
if we could, there is still the mismatch between the DOFIndices
in the ConstraintsContainer and the ContainerIndices used for
the vector.

Imported from SVN: r1827 at branches/new_dof_handling