[ISTL Backends] Use int for indexing into ContainerIndex to avoid unsigned underflow in access functions
When accessing an ISTL container with a block size
of 1 at the lowest level, we optimize the ContainerIndex
by not storing the redundant access information (which
is always 0). That causes the index into the ContainerIndex
during the recursive element access algorithms to become
negative, causing an underflow. Fixed by switching to a
signed type for that index.

Assertions on that index updated to check for the correct
value as value (0 or -1, depending on the leaf block size).

Imported from SVN: r1838 at branches/new_dof_handling