[VTK] New mechanism to automatically add all components of a GFS tree to a VTKWriter
The current way of adding discrete grid functions
to a VTKWriter is rather inconvenient, as it requires
manually creating a huge amount of GridFunctionSubSpaces
(for all non-root nodes of the GFS tree) and using those
to create DiscreteGridFunctions for the leaf spaces.

Most often, users are interested in *all* components
of the GFS tree anyway, so providing an automatic
mechanism to do so seems very useful.

This patch introduces exactly that by way of the
add_solution_to_vtk_writer(VTKWriter,GFS,X) function,
which will add all data from GFS to the VTKWriter.
If the different function spaces have been named,
those names will be used for the individual components,
otherwise the function will construct a simple name
from the component index within the GFS tree.

It is also possible to output vector-valued data
by simply replacing the PowerGridFunctionSpace of
that data by a VectorGridFunctionSpace.

Imported from SVN: r1854 at branches/new_dof_handling