[Constraints] Add do-nothing specializations of constraints-handling code
The EmptyTransformation type lacks a number of features
that are required to make constrain_residual() and the
Dirichlet constraint handling code in LocalAssemblerBase
compile correctly in all cases, so we need to circumvent
that problem by specializing the two relevant functions
on EmptyTransformation.

While doing that, the patch also clears up the matrix-
handling code in general.

Imported from SVN: r1860 at branches/new_dof_handling