[VTK] Extend new VTK output mechanism to support user-supplied output functions
It is now possible to provide custom functions working
on top of the new VTK output mechanism. Due to the way
the mechanism works, the user actually has to provide
a factory instead of the function proper, the mechanism
will then insert the actual LFS and the data holder.

This clearly needs some improvement...

Imported from SVN: r1863 at branches/new_dof_handling