[Benchmarks] Add new benchmark based on testlaplacedirichletccfv
This patch adds a new benchmark called laplacedirichletccfv, which
basically performs the same simulation as testlaplacedirichletccfv
 from the test/ directory.

This benchmark can be used to measure the performance overhead
of the new two-stage DOF -> LA index mapping infrastructure, as
the simple FV structure of the code is a worst-case scenario for the
index mapping component.
For comparison, the code can run on a second, highly simplified
infrastructure that is very similar to the old GridFunctionStaticSize
optimization, but retains things like the LFSIndexCache and the
local container views.

Other notable differences to the underlying version from test/ are:
- generates timing information for all parts of the simulation
  process
- the actual solve can be deactivated to measure raw assembly
  performance for large problems
- re-runs every simulation several times to improve timings
- individual grid / backend combinations can be enabled or disabled
- configurable through a parameter file

Imported from SVN: r1932 at branches/new_dof_handling