[Benchmarks] Add new benchmark based on testlaplacedirichletccfv
This patch adds a new benchmark called dirichletccfv, which basically
performs the same simulation as testlaplacedirichletccfv from the test/
directory.

The benchmark provides a comparison between the different mapper
strategies for the leaf GridFunctionSpace, i.e. GridFunctionGeneralMapper,
GridFunctionRestrictedMapper and SimpleGridFunctionStaticSize.

This is the version of the benchmark based on the legacy
infrastructure.

The main differences are:
- generates timing information for all parts of the simulation
process
- the actual solve can be deactivated to measure raw assembly
performance for large problems
- re-runs every simulation several times to improve timings
- individual dimension / mapper combinations can be enabled
  or disabled
- configurable through a parameter file

Imported from SVN: r1935 at trunk