[TypeTree] Add support function for unpacking function calls and ignoring their return values
Turns out the wording of the new standard only allows template argument packs to be
unpacked where the language requires a comma as an actual argument separator; this makes
it impossible to expand a pack into a statement (in this case the commas become comma
operators).

For per-element expressions that yield a return value, this problem can simply be overcome by
surrounding the unpacked statement with a call to a dummy function, as provided by the
new function discard() in this patch.

Note that in order to make this work for void expressions, it is necessary to artificially inject
a return value - ironically, this is best done with a comma operator like this:

discard((args.foo(),0)...);

The per-element expressions assume the value of the last subexpression (in this case, a simple
0).

Imported from SVN: r2009 at branches/new_dof_handling