[LocalFunctionSpace] Avoid instantiating child space orderings in LocalFunctionSpace constructor
The constructor of LocalFunctionSpace always allocates storage for the DOFIndices.
This has to major drawbacks:

- it over-allocates storage (in the end, only the storage of the root space is used)
- more importantly, it instantiates all Orderings, which the whole new API was designed
  to avoid to reduce compile times and ordering setup overhead.

This patch fixes the problem by postponing the allocation of the DOFIndex storage from
the LocalFunctionSpaceNode constructor to the setup() method, which will only be called
for the root space. As the associated root GridFunctionSpace requires an Ordering anyway,
this is unproblematic.

Imported from SVN: r2018 at branches/new_dof_handling