[Newton] Use GridOperator-style interface for determining the matrix type
Currently, the Newton solver still extracts the matrix type from the GridOperator
using the nested MatrixContainer template. This is not supported by the GridOperator
anymore (and is redundant anyway, since the GridOperator allows for explicit specification
of the matrix field type).

Note that this patch breaks compatibility with the old GridOperatorSpace.

Imported from SVN: r2042 at branches/new_dof_handling