[Stokes] Fix jacobian assembly in CG Navier-Stokes operator
In the jacobian assembly of the CG Navier-Stokes operator,
the transposed gradient part was assembled incorrectly (the
derivative was iterated over in the ansatz function and fixed in
the test function, while it has to be vice versa).

This only affects the jacobian, as the residual was already correct.

Imported from SVN: r2056 at branches/new_dof_handling