[VTK] Componentize name generation of VTK solution components
Currently, it is only possible to influence the naming of the
VTK functions generated from a solution by either naming the child
spaces or by providing a base name, which will be combined with
the tree paths of the children to generate unique function names.

Unfortunately, this naming scheme fails when trying to add two different
solutions to a single VTK file with named sub spaces, as there is no way
to generate different names for the component functions.

This patch turns the name generation of the VTK output mechanism
into a component that is passed in by the user and gets called by the
mechanism with the GFS name and its associated TreePath.

There is also a default implementation that can be obtained from the
function default_vtk_name_scheme and which allows setting a custom
prefix, suffix and TreePath item separator.

Imported from SVN: r2057 at branches/new_dof_handling