[LocalFunctionSpace] Fix LocalFunctionSpace for GridFunctionSpaces with intersection DOFs
This patch updates the method multiIndices() in LocalFunctionSpace to correctly handle
intersection DOFs. In order to do so, the LFS needs to delegate the index calculation to
the IntersectionIndexSet (IIS) of the underlying GFS.Unfortunately, the IIS does not exist
for most GFS instances, and for the specialization where it exists, it is not exported.

In order to solve those issues, we

1) export the type of the IntersectionIndexSet in the relevant specialization and provide
   a method intersectionIndexSet() to access it.

2) do some SFINAE magic to switch between accessing the IIS for intersection DOFs if
   it is available or falling back to throwing an exception if it isn't.

Patch compiles, but is not runtime-tested, as ALUGrid is acting up on me...

Imported from SVN: r2064 at trunk