[OneStepMethod] Store a reference to the solver instead of a copy
The OneStepMethod copies the PDESolver object it gets passed in the constructor,
which is both inconsistent with the usual PDELab API (which tends to pass by
(const) reference) and also makes it impossible to tune solver parameters after
creating the OneStepMethod.

This patch switches to storing a *mutable* reference to the PDESolver (required
as the OneStepMethod needs to call some non-const methods on the PDESolver).
Note that this change introduces two subtle API changes:

- The OneStepMethod constructor now takes the PDESolver by mutable reference,
  not by const reference.
- Changing the PDESolver object after creating the OneStepMethod *will now change*
  the behavior of the OneStepMethod (well, that's part of the point of this patch...).

Patch by Rebecca Neumann.

Imported from SVN: r2126 at trunk