[LocalFunctionSpace] Make sure the _global pointer is correctly initialized during root LFS construction
Due to the way LFS construction works, the default constructor for the interface class
LocalFunctionSpace inherits from the acutal LocalFunctionSpaceNode corresponding to its
underlying GFS and constructs this base class using the TypeTree transformation mechanism.
This process involves copy-constructing the actual base class instance from the return value
of the transform() call and subsequently fixing up all the _global pointers in the LFS
hierarchy to point to the root _global_storage container. Unfortunately, we forgot to adjust
the pointer of the root space before calling setup, causing it to point to the _global_storage
of the temporary returned by the TypeTree transformation.

For some weird reason, this bug did not show up in any way in binaries produced by GCC (even
valgrind did not throw any kind of warning), while clang-generated binaries immediately tripped
over it.

This essentially fixes the same problem r2082 solves in the branch with the new DOF handling.

Imported from SVN: r2127 at trunk