[ISTLMatrixBackend] Allow addressing entries in Accessor using global indices
This patch adds methods setGlobal(), addGlobal() and getGlobal() to the accessor.
These methods are identical to their existing counterparts, but expect the indices
of the matrix entry to already be mapped into the global index space.

This is in preparation for reenabling support for general constraints.

Imported from SVN: r2128 at trunk