[PETScMatrixBackend] Provide stubs for globally addressed API in Accessor
This patch adds methods setGlobal(), addGlobal() and getGlobal() to the accessor.
These methods all raise Dune::NotImplemented, as we cannot guarantee that the
passed-in entry is part of the matrix view represented by the accessor. Support
for this feature will be implemented as part of the new DOF handling.

This is in preparation for reenabling support for general constraints.

Imported from SVN: r2130 at trunk