[LocalAssemblerBase] Reenable support for general constraints
During the introduction of the PETSc backend and the switch to
Accessor-based matrix access in the assembler, the support for
general constraints was broken.

This patch reenables that support. As general constraints can
grow the local matrix stencil and extracting this information is
quite a bit of work, general constraints are not supported for the
PETSc backend (which needs to know the exact local matrix stencil).
Support for general constraints with PETSc will come as part of the
new DOF handling after the 1.0 release.

The implementation does a little bit of a dance to figure out whether
the current matrix entry to be written belongs to a general constraint
and uses the newly introduced Accessor API for writing entries based
on their global index in this case. For all other entries, the code
continues to use the access methods based on local indices, retaining
PETSc compatibility for unconstrained DOFs and those with Dirichlet
constraints.

Imported from SVN: r2131 at trunk