[Ordering] Rewrite of per-codim flag handling and new flag for fixed size per codim
The current Ordering implementation stores per-codim data in std::vector<bool>, which
seems like a bit of overkill (just the start pointer in the vector uses more space
than we need for our few codimensions...), so this patch moves those flags into a
std::bitset. In order to have a central location for changing the maximum supported
codim, the OrderingTraits export a new type CodimFlag that is used everywhere per-codim
flags need to be stored.

Apart from that, the patch also introduces a new flag that tracks whether codims are
fixed size. This is preparatory work for the DataHandle support (data handles need access
to this information). Right now, the implementation is not overly clever, it either claims
that either no codim is of fixed size or all of them, but that may be improved later on.