[DataHandle] New mixin class DataHandleProvider for adding DataHandle support to GridFunctionSpaces
This patch adds a new class DataHandleProvider that will be used as a mixin for the GridFunctionSpaces and provides the methods required for DataHandle support by the DataHandle API. As the actual calculation was moved from the GridFunctionSpace to the Ordering, we can use the same implementation for all GridFunctionSpaces, so it is factored out to the mixin to avoid code duplication.

The main change w.r.t. the old interface is the method dataHandleContainerIndices(), which obviously operates on container indices instead of traditional global indices now. Apart from that, the method now also returns the number of indices associated with the entity, as this information needs to be calculated anyway, and by returning that information, we avoid doing the calculation twice.