[LFSIndexCache] Provide typedefs for DOFIndex and ContainerIndex
This patch adds typedefs for the index types to the LFSIndexCache. Currently, code has to
extract this information from the LocalFunctionSpace typedef, but that creates problems with
other kinds of index caches that are not based on a LocalFunctionSpace. We will need such
an index cache for the generic data handle.