[Constraints] Rewrite constraints to be based on ContainerIndex instead of DOFIndex
During the initial conversion of the constraints to the new index framework, it wasn't
really clear whether to use DOFIndices or ContainerIndices for the constraints container.
I considered DOFIndices to be a more natural choice, as the constraints are fundamentally
a property of the GridFunctionSpace and as such, should be based on the indices it exports.
On the other hand, as the constraints will always be applied to vectors / matrices, having
ContainerIndices greatly simplifies constraints application.

With the parts of PDELab converted up to now, constraints application has always happened
at the level of the DOFs associated with a LocalFunctionSpace, and in that context, having
the constraints in terms of DOFIndices has worked reasonably well. But during the conversion
of the parallel ISTL solver support, it turned out that it is often necessary to apply constraints
at the container level in a single operation, something which is *a lot* easier when the constraints
are already expressed in the right kind of index.

So this patch completely tears apart the constraints handling code and switches it over to
ContainerIndices. Moreover, the constraints assembler now works in a way that more closely
resembles the other assemblers, i.e. it collects all constraints associated with a LocalFunctionSpace
before inserting them into the global constraints container, whereas the old code inserted the
data after each visit to a leaf space.

The changes in detail:

- The ConstraintsTransformation now takes three template parameters(DOFIndex, ContainerIndex and FieldType).
- The LocalTransformation is now indexed by the DOFIndex, not the (very inconsistent) leaf-local index.
  This requires a change in the constraints engines, which is probably a good thing as unconverted container
  accesses will trigger a compiler error and make sure people notice that their code needs to be adapted.
  This change also allows for more general constraints, as it now becomes possible to have constraints
  based on arbitrary (non-local) degrees of freedom.
- The constraints assembler now collects all constraints for a single LocalFunctionSpace in a LocalTransformation
  before adding it to the global transformation. For this import process, the global transformation offers
  a new method import_local_transformation(). I did not call it add_local_transformation(), as it will simply
  overwrite existing column entries (this behavior is unchanged from the old implementation).
- The existing constraints engines (apart from the hanging nodes, which require more work) are ported to be
  compatible with the new-style LocalTransformation.
- Having the constraints specified using ContainerIndices reduces the amount of work to be done when updating
  the LFSContainerIndexCache, as we don't have to map the constraints anymore at bind time.
- The diverse utility functions in constraints.hh to set / copy constrained / unconstrained DOFs, which were
  changed to require the GridFunctionSpace as part of the initial porting effort are changed back to their
  initial signatures (i.e. you don't need to pass the GridFunctionSpace anymore), as we don't need to map
  the indices of the constraints anymore.