[DataHandle] Make data writeback dependent on GatherScatter return value
The GFSDataHandle now expects the GatherScatter functor to return a bool
value from its gather() and scatter() methods and uses the return value
to decide whether or not the LocalView needs to be committed to the
global container. Note that this is just a performance optimization, allowing
certain backends to avoid the overhead associated with performing the data
writeback. This is no transaction in the database sense and there is NO
guarantee that data will not be written if gather() or scatter() return false,
as e.g. ISTL always writes directly into the global container.

The reason for this change is that most of the time, data will only be written
on scatter(), but some of the ParallelISTLHelper support actually requires
writing data at the gather() stage, and I wanted to avoid having to pay the
price of ensuring data writeback for each GatherScatter implementation.

This patch also changes the pre-made DataHandles in genericdatahandle.hh to
always return false on gather() and true on scatter().