[ISTLVectorBackend] Provide new method dot()
This patch adds a method dot() to the ISTLBlockVectorContainer that
simply delegates to the dot() method of the underlying ISTL vector.
This change is a stopgap measure to provide compatibility with the
recently introduced difference between operator*() and dot() in ISTL
and Common.

In the long run, we should probably adapt the backends to only call
ISTL routines on the contained ISTL objects.

Imported from SVN: r2159 at trunk