[GridViewOrdering] Fix blocked container handling for scalar GridViewOrdering
The scalar version of the GridViewOrdering always sets size == block_count, which
is obviously wrong for blocked containers. This patch fixes the issue for fixed
size spaces. Variably-sized spaces require more work, and I'm not entirely sure
if it's worth the effort (we would only be able to have the per-entity size of the
space vary in multiples of the block size). The patch adds a warning anyway.