Adapt cg_stokes.hh to new stokesparameter.hh interface.
***** HACK ALARM *******
I had to change the interface of
TaylorHoodNavierStokesDefaultParameters::rho(eg,x). Namely the type of
x is not fixed anymore, but can be arbitrary. This was needed as there
is no way of finding out the global coordinate within
NavierStokesMass::jacobian_volume (or at least it is unclear which
value to take).


Imported from SVN: r2233 at trunk