[GlobalDOFIndex] Add data structure for global DOF index based on entity id
When communicating DOFs over MPI, we cannot use the standard DOFIndex as it is based
on local entity identifiers (GridView + GeometryType + index).

This patch resolves the problem by introducing a GlobalDOFIndex, replacing the entity
information in the standard DOFIndex with the (global) ID of the entity. Note that
like the standard DOFIndex, this type is hashable, but not (yet) ordered, so it can
only be used in conjunction with unordered_{set,map}.
This creates a bit of a problem, as the IDs of ALUGrid and YaspGrid are currently not
hashable. I have submitted patches to dune-grid to resolve that issue.