[OPBLocalFiniteElement] Fix multiple linkage errors of some global functions
The implementation of OPBLocalFiniteElement contains some global functions that
are not templated and can thus create multiple linkage errors.

Solve the problem by declaring the functions inline. An alternative would be to
move them to a separate object, but that would make inlining impossible, and
they aren't really big.

Thanks to Matthias Wohlmuth for the patch.

Imported from SVN: r2266 at trunk