[ISTL] Create subdirectory for ISTL implementation artifacts and add container tagging infrastructure
With the growing number of files in the backend directory, I decided to add
backend-specific subdirectories for implementation-specific details of the
diverse backends, so that only the main entry headers should stay in the
backend directory in the long run.

As a first of those implementation helpers, this patch also introduces an
infrastructure for tagging ISTL and Field* containers to simplify writing
the recursive algorithms required for handling them.