[GridFunctionSpace] Add new ordering tags for leaf spaces
The old SizeTag, which was used to partially specialize the leaf
GridFunctionSpace according to the required flexibility w.r.t.
to per-entity DOF counts is not needed anymore, as those algorithms
have moved into the orderings. On the other hand, it may be useful
to pass some policies to the generated ordering.

So this patch introduces a new LeafOrderingTag that takes a single
parameter object from which it inherits, allowing injection of
additional types and constants into the tag.

For now, there are implementations, a default one and one for non-
overlapping grids, where the ordering must never use a fixed-size
mapping to correctly handle the ghost layer.